Cast ptr subtraction.  From Fouts.

+++ gpsbabel/csv_util.c
@@ -102,7 +102,7 @@ csv_stringtrim(const char *string, const char *enclosure)
 
     /* if we have enclosures, skip past them in pairs */
     if (elen) {
-	while (((p2 - p1) > elen) &&
+	while (((size_t) (p2 - p1) > elen) &&
 	       (strncmp(p1, enclosure, elen) == 0) &&
 	       (strncmp((p2 - elen + 1), enclosure, elen) == 0)) {
 	    p2 -= elen;
